Update
This update is a fix to make the WaterPiston work there is still much
more to be done to make it more effecent but that will be in future
versions.  Also I have added the GardenHarvester-UserMenu to the code
this is what is used to manage the harvesting of the Garden.  It too
will be managed and updated to take less time and be more effecent along
with giving more options as right now once you are in it wont let you
exit.