New Code
I have uploaded the First of many to come Turtle codes.  This one allows
you to harvest Fir Tree's this will require a setup of a turtle station
(as I call it) to allow the turtles to refill and dump their collected
items along with gathering more suppies to keep running.